Sketched out 'physics' package.
