+++ game-engine/src/main/java/gameengine/engine/IScene.java
@@ -3,7 +3,7 @@ package gameengine.engine;
 import gameengine.engine.renderer.IRenderable;
 import gameengine.engine.renderer.component.Camera;
 
-public interface IScene extends IRenderable {
+public interface IScene extends IRenderable, ITickable {
 	
 	public void setActiveCamera(Camera camera);
