+++ game-engine/src/main/java/gameengine/engine/physics/IPhysicsObject.java
@@ -1,0 +1,6 @@
+package gameengine.engine.physics;
+
+public interface IPhysicsObject extends IHasTransform {
+
+	public Physics getPhysics();
+}
