+++ game-engine/src/main/java/gameengine/game/Scene.java
@@ -7,20 +7,23 @@ import org.joml.Vector3f;
 
 import gameengine.engine.IGameObject;
 import gameengine.engine.IScene;
-import gameengine.engine.ITickable;
+import gameengine.engine.physics.IPhysicsObject;
+import gameengine.engine.physics.PhysicsScene;
 import gameengine.engine.renderer.Renderer;
 import gameengine.engine.renderer.component.Camera;
 import gameengine.game.component.light.AmbientLight;
 import gameengine.game.component.light.DirectionalLight;
 
-public class Scene implements IScene, ITickable {
+public class Scene implements IScene {
 	private List<IGameObject> objects;
+	private PhysicsScene physicsScene;
 	private Camera activeCamera;
 	private AmbientLight ambientLight;
 	private DirectionalLight directionalLight;
 	
 	public Scene() {
 		this.objects = new ArrayList<>();
+		this.physicsScene = new PhysicsScene();
 		this.activeCamera = null;
 		this.ambientLight = new AmbientLight(new Vector3f(1, 1, 1), 0.5f);
 		this.directionalLight = new DirectionalLight(new Vector3f(1, 1, 1), 1.0f);
@@ -48,6 +51,11 @@ public class Scene implements IScene, ITickable {
 	public void addObject(IGameObject object) {
 		this.objects.add(object);
 		object.onCreate();
+	}
+	
+	public void addObject(IPhysicsObject object) {
+		this.addObject(object);
+		this.physicsScene
 	}
 	
 	public void setActiveCamera(Camera camera) {
