+++ game-engine/src/main/java/gameengine/test/TestModel.java
@@ -1,10 +1,10 @@
 package gameengine.test;
 
 import gameengine.engine.IGameObject;
+import gameengine.engine.physics.IHasTransform;
+import gameengine.engine.physics.Transform;
 import gameengine.engine.renderer.Renderer;
-import gameengine.game.component.IHasTransform;
 import gameengine.game.component.Model;
-import gameengine.game.component.Transform;
 
 public class TestModel implements IGameObject, IHasTransform {
